consolidate into CollectionService
